Implement advanced copying while resizing
After creating a new array elements are copied according to their position in the array. If the indexOfLast is more or equal to indexOfFirst, then (indexOfLast - indexOfFirst + 1) elements will be copied in the destination array. If indexOfLast is less than indexOfFirst, then (initialCapacity - indexOfFirst) elements will be copied to the beginning of the new array, and (indexOfLast + 1) elements will be copied to the end of the new array with update of the indexOfFirst and indexOfLast.