Add updating indexes after removing all elements
If there is only one element left in the deque, its indexOfFirst and indexOfLast are equal. Which means that any subsequent call of the popFirst() or popLast() means that the last element was removed. If the last element was removed, it is allowed to add elements starting from the 0th index. So any call of popFirst() when indexOfFirst and indexOfLast are equal will cause reset of the indexOfFirst and indexOfLast to -1.